% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{deparse}
\alias{deparse}
\alias{deparsec}
\title{A nicer deparse}
\usage{
deparse(x, ...)

deparsec(x, ...)
}
\description{
\code{deparse} is a reimplementation of \code{\link[base]{dput}} and related
functions. It tries its best to produce output that is easy to read
(for humans), yet produces (almost) identical results to the input
(for machines). This function is a generic, so other packages can easily
provide implementations for the objects they define.

The \code{deparsec} function leverages \code{deparse} by creating
a \code{call} object which can be evaluated but retains formatting
(in the form of a \code{\link[base]{srcref}} attribute).
}
