% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp}
\alias{dp}
\alias{dpc}
\title{A nicer deparse}
\usage{
dp(x, ...)

dpc(x, ...)
}
\description{
\code{dp} is a reimplementation of \code{\link[base]{dput}} and related
functions. It tries its best to produce output that is easy to read
(for humans), yet produces (almost) identical results to the input
(for machines). This function is a generic, so other packages can easily
provide implementations for the objects they define.

The \code{dpc} function leverages \code{dp} by creating
a \code{call} object which can be evaluated but retains formatting
(in the form of a \code{\link[base]{srcref}} attribute).
}

