% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{lave}
\alias{lave}
\alias{lavec}
\title{A nicer deparse}
\usage{
lave(x, ...)

lavec(x, ...)
}
\description{
\code{lave} is a reimplementation of \code{\link[base]{dput}} and related
functions. It tries its best to produce output that is easy to read
(for humans), yet produces (almost) identical results to the input
(for machines). This function is a generic, so other packages can easily
provide implementations for the objects they define.

The \code{lavec} function leverages \code{lave} by creating
a \code{call} object which can be evaluated but retains formatting
(in the form of a \code{\link[base]{srcref}} attribute).
}

