% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcref-call.R
\name{srcref_call}
\alias{srcref_call}
\alias{as.srcref_call}
\title{Calls with attached srcrefs}
\usage{
srcref_call(call, srcref)

as.srcref_call(x, ...)
}
\description{
As of R 2.10.0, \code{\link[base]{call}} objects don't use a
\code{\link[base]{srcref}} attribute
even if it's attached, see \code{vignette("calls", package = "deparse")} for
details. This class provides call objects that use their \code{srcref}
attribute for printing.
}
