% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp}
\alias{dp}
\title{A nicer deparse}
\usage{
dp(x, ...)
}
\description{
\code{dp} is a reimplementation of \code{\link[base]{dput}} and related
functions. It tries its best to produce output that is easy to read
(for humans), yet produces (almost) identical results to the input
(for machines).

Unlike \code{\link[base]{dput}} and \code{\link[base]{deparse}}, it creates
a \code{call} object which can be evaluated.  Because the formatting may
be important, it is retained as an attribute.
}

